Import Boost 1.69.0

+++ boost/log/support/date_time.hpp
@@ -22,7 +22,7 @@
 #include <iterator>
 #include <boost/cstdint.hpp>
 #include <boost/move/core.hpp>
-#include <boost/move/utility.hpp>
+#include <boost/move/utility_core.hpp>
 #include <boost/date_time/time.hpp>
 #include <boost/date_time/date.hpp>
 #include <boost/date_time/gregorian/gregorian_types.hpp>
@@ -55,17 +55,17 @@ inline void decompose_date(DateT const& d, boost::log::aux::decomposed_time_wrap
 {
     typedef typename DateT::ymd_type ymd_type;
     ymd_type ymd = d.year_month_day();
-    v.year = ymd.year;
-    v.month = ymd.month;
-    v.day = ymd.day;
+    v.year = static_cast< uint32_t >(ymd.year);
+    v.month = static_cast< uint32_t >(ymd.month);
+    v.day = static_cast< uint32_t >(ymd.day);
 }
 
 template< typename TimeDurationT, typename ValueT >
 inline void decompose_time_of_day(TimeDurationT const& tod, boost::log::aux::decomposed_time_wrapper< ValueT >& v)
 {
-    v.hours = tod.hours();
-    v.minutes = tod.minutes();
-    v.seconds = tod.seconds();
+    v.hours = static_cast< uint32_t >(tod.hours());
+    v.minutes = static_cast< uint32_t >(tod.minutes());
+    v.seconds = static_cast< uint32_t >(tod.seconds());
 
     typedef typename TimeDurationT::traits_type traits_type;
     enum
@@ -96,10 +96,10 @@ inline void decompose_date_duration(DateDurationT const& dur, boost::log::aux::d
     if (dur.is_negative())
     {
         v.negative = true;
-        v.day = (-dur).days();
+        v.day = static_cast< uint32_t >((-dur).days());
     }
     else
-        v.day = dur.days();
+        v.day = static_cast< uint32_t >(dur.days());
 }
 
 template< typename TimeT, typename ValueT >
